[housekeeping] refactor codebase to be modular
