+++ src/handlers/protected.rs
@@ -1,0 +1,5 @@
+use axum::Json;
+
+pub async fn protected_handler() -> Json<&'static str> {
+    Json("Access granted!")
+}
