+++ src/types/mod.rs
@@ -1,0 +1,5 @@
+pub mod channel;
+pub mod error;
+
+pub use channel::{PaymentChannel, SignedRequest};
+pub use error::AuthError;
